BC break, but it's a Google beta, so.
Segments now lazy-load themselves.  On read, they will reactivate an available session, but will not start a new one.  On writes, they will reactivate an available session, or start a new one if one is not available.

This means that creating a segment object no longer starts a session; you have to read from or write to one for the session to start.